chore: move examples into node package
