+++ packages/elgato-stream-deck/examples/brightness.js
@@ -1,4 +1,4 @@
-const { openStreamDeck } = require('../packages/elgato-stream-deck-core/dist/index')
+const { openStreamDeck } = require('../dist/index')
 const streamDeck = openStreamDeck()
 
 // Fill it white so we can see the brightness changes
