+++ packages/elgato-stream-deck/examples/device-detection.js
@@ -1,5 +1,5 @@
 const usbDetect = require('usb-detection')
-const { listStreamDecks, openStreamDeck } = require('../packages/elgato-stream-deck-core/dist/index')
+const { listStreamDecks, openStreamDeck } = require('../dist/index')
 const streamDecks = {}
 
 function addDevice(info) {
