+++ packages/elgato-stream-deck/examples/fixtures/github_logo.png
