+++ packages/elgato-stream-deck/examples/fixtures/sunny_field.png
