+++ packages/elgato-stream-deck/examples/text-generation.js
@@ -2,7 +2,7 @@ const sharp = require('sharp')
 const path = require('path')
 const PImage = require('pureimage')
 const streamBuffers = require('stream-buffers')
-const { openStreamDeck } = require('../packages/elgato-stream-deck-core/dist/index')
+const { openStreamDeck } = require('../dist/index')
 
 const streamDeck = openStreamDeck()
 streamDeck.clearAllKeys()
